Add files via upload
Using the reinforcement learning method, it is necessary to teach the cart to move away from the wall to the maximum distance. The reward is presented as the value of the change in the distance from the wall to the cart during movement. Distance D from the wall is measured with a rangefinder. Movement in this example is possible only with a certain displacement of the “drive”, consisting of two arrows S1 and S2. The arrows are two servos with guides connected in the form of a “knee”. Each servo in this example can turn 6 equal angles. The model has the ability to perform 4 actions, which are the control of two servos, actions 0 and 1 rotate the first servo by a certain angle clockwise and counterclockwise, actions 2 and 3 rotate the second servo by a certain angle clockwise and counterclockwise. Figure 1 shows a working prototype car.
Here we used two learning algorithms (SARSA ad Q-Learning) with different values of the hyperparameters.